Add ability to run gekko in a docker container using docker-compose.
Rationale: Adding docker-compose makes it easier for contributers and
end-users to interact with their docker containers. For example,
setting up and running gekko using docker-compose could be as easy as
`docker-compose up gekko`. Simply starting, stopping or restarting it as
a daemon could be `docker-compose start`, `docker-compose stop`, or
`docker-compose restart` respectively.